requirements.txt
python_version<=3.10.9
opencv-python
moms_apriltag
pillow
pillow_heif
numpy
matplotlib
albumentations